% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_paths.R
\name{get_file_paths}
\alias{get_file_paths}
\title{A function the returns the file paths to example emails for testing.}
\usage{
get_file_paths()
}
\value{
A vector of file paths
}
\description{
A function the returns the file paths to example emails for testing.
}
