% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emails.R
\name{read_emails}
\alias{read_emails}
\title{A function to extract To, From, Date, Subject, and Message fields from a  
collection of emails saved as plain text.}
\usage{
read_emails(email_files)
}
\arguments{
\item{email_files}{A character vector containing the full paths to the email 
files to be processed.}
}
\value{
An data.frame with the following columns: to, from, CC, subject, 
date, message, message_tokenized, subject_tokenized, num_recipients, 
all_tokens, num_tokens
}
\description{
A function to extract To, From, Date, Subject, and Message fields from a  
collection of emails saved as plain text.
}
\examples{
files <- get_file_paths()
emails <- read_emails(files)
}
