% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Read_In_Email.R
\name{Read_In_Email}
\alias{Read_In_Email}
\title{A function to extract To, From, Date, Subject, and Message fields from a plain text email.}
\usage{
Read_In_Email(email_file)
}
\arguments{
\item{email_file}{The full path to the email file.}
}
\value{
An object of class Email that contains the following members: to, from, CC, subject, date, message, message_tokenized, subject_tokenized, num_recipients, all_tokens, num_tokens
}
\description{
A function to extract To, From, Date, Subject, and Message fields from a plain text email.
}
\examples{
MyEmail <- Read_In_Email("/Users/matthewjdenny/Dropbox/PINLab/Raw_Emails/2")
}

