\name{REmail-package}
\alias{REmail-package}
\alias{REmail}
\docType{package}
\title{
\packageTitle{REmail}
}
\description{
\packageDescription{REmail}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{REmail}
\packageIndices{REmail}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{REmail}

Maintainer: \packageMaintainer{REmail}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
