\name{REmail-package}
\alias{REmail-package}
\alias{REmail}
\docType{package}
\title{
\packageTitle{REmail}
}
\description{
\packageDescription{REmail}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{REmail}
\packageIndices{REmail}
This package is still in development
}
\author{
\packageAuthor{REmail}

Maintainer: \packageMaintainer{REmail}
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
Read_In_Email("Myemail")
}
